Move code around and adjust public/private
